refactored splash to extract single run checks
